+++ jaot_base/tests/__init__.py
@@ -1,5 +1,6 @@
 from . import test_data
 from . import test_formulations
+from . import test_presentation
 from . import test_models
 from . import test_lifecycle
 from . import test_security