+++ jaot_mrp/tests/__init__.py
@@ -2,3 +2,4 @@
 # License LGPL-3
 from . import test_mrp_formulation
 from . import test_mrp_e2e
+from . import test_presentation